+++ app/pb_kits/playbook/pb_hashtag/docs/_hashtag_dark.jsx
@@ -20,6 +20,13 @@ const HashtagDark = () => {
       <br />
       <Hashtag
           dark
+          text="456789"
+          type="appointment"
+          url="https://google.com"
+      />
+      <br />
+      <Hashtag
+          dark
           text="654321"
           type="default"
           url="https://google.com"